+++ routers/fmapi.js
@@ -31,13 +31,27 @@ router.use(function (req, res, next) {
 });
 
 router.get("/scanDirectory", function (req, res) {
-  fmapi.scanDirectory(req.query.server, req.query.directory, function (data) {
-    res.send(data);
-  });
+  if (
+    additional.validatePath(req.query.directory) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    fmapi.scanDirectory(req.query.server, req.query.directory, function (data) {
+      res.send(data);
+    });
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/getFile", function (req, res) {
-  res.send(fmapi.readFile(req.query.server, req.query.path));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    res.send(fmapi.readFile(req.query.server, req.query.path));
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/packetRemoving", function (req, res) {
@@ -72,30 +86,78 @@ router.get("/packetRemoving", function (req, res) {
 });
 
 router.get("/saveFile", function (req, res) {
-  res.send(fmapi.saveFile(req.query.server, req.query.path, req.query.text));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    res.send(fmapi.saveFile(req.query.server, req.query.path, req.query.text));
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/downloadFile", function (req, res) {
-  path_download = path.resolve(
-    "./servers/" + req.query.server + req.query.path
-  );
-  res.download(path_download);
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    path_download = path.resolve(
+      "./servers/" + req.query.server + req.query.path
+    );
+    res.download(path_download);
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/deleteFile", function (req, res) {
-  res.send(fmapi.deleteFM(req.query.server, req.query.path));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    res.send(fmapi.deleteFM(req.query.server, req.query.path));
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/deleteDirectory", function (req, res) {
-  res.send(fmapi.deleteDirFM(req.query.server, req.query.path));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true
+  ) {
+    res.send(fmapi.deleteDirFM(req.query.server, req.query.path));
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/renameFile", function (req, res) {
-  res.send(fmapi.renameFM(req.query.server, req.query.path, req.query.newname));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true &&
+    additional.validatePath(req.query.newname) == true
+  ) {
+    res.send(
+      fmapi.renameFM(req.query.server, req.query.path, req.query.newname)
+    );
+  } else {
+    res.status(403).send();
+  }
 });
 
 router.get("/newDirectory", function (req, res) {
-  res.send(fmapi.newdirFM(req.query.server, req.query.path, req.query.newdir));
+  if (
+    additional.validatePath(req.query.path) == true &&
+    additional.validatePath(req.query.server) == true &&
+    additional.validatePath(req.query.newdir) == true
+  ) {
+    res.send(
+      fmapi.newdirFM(req.query.server, req.query.path, req.query.newdir)
+    );
+  } else {
+    res.status(403).send();
+  }
 });
 
 module.exports = router;
